+++ packages/app/src/components/puzzle/Victory.vue
@@ -11,7 +11,7 @@ import {
   BRONZE,
   NOT_RANKED,
   levelByMoves
-} from "../../../core/src/levels";
+} from "../../../../core/src/levels";
 
 const victorySentences = {
   [GOLD]: "Congrats, you are now the 15 Puzzle master!",
