+++ packages/app/src/layout/Content.vue
@@ -1,12 +1,12 @@
 
 <template>
-  <div id="content">
+  <main id="content">
     <Game/>
-  </div>
+  </main>
 </template>
 
 <script>
-import Game from "../components/Game.vue";
+import Game from "../components/puzzle/Game.vue";
 
 export default {
   name: "Content",
@@ -27,5 +27,12 @@ export default {
   height: 100%;
   min-height: 100%;
   margin-top: 80px;
+  padding: 5px;
+}
+
+@media screen and (min-width: 600px) {
+  #content {
+    padding: 5px 5%;
+  }
 }
 </style>
